Remove ExampleExtractor and simplify code architecture
# Changes by File

## lib/loco_motion/algolia/example_extractor.rb
- Removed file completely as it was an unnecessary abstraction
- Moved functionality into SearchRecordBuilder for simplification

## lib/loco_motion/algolia/search_record_builder.rb
- Added private extract_examples method with core functionality from ExampleExtractor
- Removed dependency on ExampleExtractor
- Simplified the enhanced_examples output by removing unused file_path

## lib/tasks/algolia.rake
- Removed require statement for example_extractor
- Reordered remaining require statements for better organization
